requirements.txt
numpy
Pillow
midvoxio
trimesh